created a new file
In this project, we will create a homepage that will display a logo and some menu items. we can also have some fun with the keyframes tag for animations. Note how the navigation menu and the rest of the page are separated using the height property. we can apply a background image, and change the height to 85vh or 100vh and see the difference it makes.